refactor: split small nbnc hotspots without def includes
